various changes to clean warnings from MSVC builds
git-svn-id: https://svn.osgeo.org/gdal/trunk@21680 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/gtiff/libgeotiff/geotiff_proj4.c
@@ -115,7 +115,7 @@ static char **OSRProj4Tokenize( const char *pszFull )
 static const char *OSR_GSV( char **papszNV, const char * pszField )
 
 {
-    int field_len = strlen(pszField);
+    size_t field_len = strlen(pszField);
     int i;
     
     if( !papszNV )
